// src/app/models/bank-transactions.model.ts
export class BankTransactionModel {
    mobileNo: string;
    accNum: string;
    accType: string;
    accIfsc: string;
    accName: string;
    bankName: string;
    bankCode: string;
    bankBranch: string;
    upiId: string;
    txnType: string;
    debit: string;
    credit: string;
    datetimeCreated: string;
    txnNarrative: string;
    txnRefNo: string;
    accBalance: string;
}